# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Streetcred streetwear e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sizes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  available: number;
  featured?: number;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  price: number;
  category?: string;
  sizes?: string;
  imageUrl?: string;
  featured?: number;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  sizes?: string;
  imageUrl?: string;
  available?: number;
  featured?: number;
}

export interface AvailabilityUpdate {
  available: number;
}

export type ProductStatsCategoriesItem = {
  category: string;
  count: number;
};

export interface ProductStats {
  total: number;
  available: number;
  unavailable: number;
  categories: ProductStatsCategoriesItem[];
}

export interface Announcement {
  id: number;
  message: string;
  active?: number;
  createdAt: string;
}

export interface AnnouncementInput {
  /** @minLength 1 */
  message: string;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminToken {
  token: string;
}

export interface ChangePasswordInput {
  currentPassword: string;
  newPassword: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface UploadResponse {
  url: string;
}

export type ListProductsParams = {
category?: string;
available?: string;
};