# lib/api-zod/src/generated/types/announcementInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Streetcred streetwear e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface AnnouncementInput {
  /** @minLength 1 */
  message: string;
}