# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Streetcred streetwear e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sizes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  available: number;
  featured?: number;
  createdAt: string;
}