# lib/api-zod/src/generated/types/productStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Streetcred streetwear e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatsCategoriesItem } from './productStatsCategoriesItem';

export interface ProductStats {
  total: number;
  available: number;
  unavailable: number;
  categories: ProductStatsCategoriesItem[];
}