# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Streetcred streetwear e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  sizes?: string;
  imageUrl?: string;
  available?: number;
  featured?: number;
}